+++ visionmetrics/classification/__init__.py
@@ -1,17 +1,18 @@
 # Import metrics directly from torchmetrics
-from torchmetrics.classification import (BinaryAUROC, MulticlassAccuracy,
-                                         MulticlassAUROC,
+from torchmetrics.classification import (BinaryAUROC, MulticlassAUROC,
                                          MulticlassAveragePrecision,
                                          MulticlassCalibrationError,
                                          MulticlassConfusionMatrix,
-                                         MulticlassExactMatch,
-                                         MulticlassPrecision, MulticlassRecall,
                                          MultilabelAccuracy, MultilabelAUROC,
                                          MultilabelAveragePrecision,
                                          MultilabelConfusionMatrix,
                                          MultilabelExactMatch)
 
 # Import custom metrics from visionmetrics
-from visionmetrics.classification.precision_recall import (MultilabelF1Score,
+from visionmetrics.classification.accuracy import MulticlassAccuracy
+from visionmetrics.classification.precision_recall import (MulticlassF1Score,
+                                                           MulticlassPrecision,
+                                                           MulticlassRecall,
+                                                           MultilabelF1Score,
                                                            MultilabelPrecision,
                                                            MultilabelRecall)
