+++ visionmetrics/classification/accuracy.py
@@ -1,0 +1,7 @@
+import torchmetrics
+from visionmetrics.classification.common import MulticlassMixin
+
+
+class MulticlassAccuracy(MulticlassMixin, torchmetrics.classification.MulticlassAccuracy):
+    pass
+
