+++ visionmetrics/classification/precision_recall.py
@@ -1,44 +1,27 @@
 import torchmetrics
-from typing import Literal
 
-from visionmetrics.prediction_filters import TopKPredictionFilter
-
-
-class MultilabelExtend:
-    """
-    Extend torchmetrics.classification.Multilabel* metrics to support:
-        1. top_k filtering
-        2. 'indices' prediction_mode
-    """
-
-    def __init__(self, top_k: int = None, prediction_mode: Literal['prob', 'indices'] = 'prob'):
-        self.top_k = top_k
-        self.prediction_mode = prediction_mode
-
-    def update(self, predictions, targets):
-        # torchmetrics only supports prob mode
-        if self.top_k is None and self.prediction_mode == 'prob':
-            super().update(predictions, targets)
-        else:
-            self.top_k = predictions.shape[1] if self.top_k is None else self.top_k
-            pred_filter = TopKPredictionFilter(k=self.top_k, prediction_mode=self.prediction_mode)
-            topk_preds = pred_filter.filter(predictions)
-            super().update(topk_preds, targets)
+from visionmetrics.classification.common import MulticlassMixin, MultilabelMixin
 
 
-class MultilabelPrecision(MultilabelExtend, torchmetrics.classification.MultilabelPrecision):
-    def __init__(self, top_k=None, prediction_mode='prob', *args, **kwargs):
-        MultilabelExtend.__init__(self, top_k=top_k, prediction_mode=prediction_mode)
-        torchmetrics.classification.MultilabelPrecision.__init__(self, *args, **kwargs)
+class MulticlassPrecision(MulticlassMixin, torchmetrics.classification.MulticlassPrecision):
+    pass
 
 
-class MultilabelRecall(MultilabelExtend, torchmetrics.classification.MultilabelRecall):
-    def __init__(self, top_k=None, prediction_mode='prob', *args, **kwargs):
-        MultilabelExtend.__init__(self, top_k=top_k, prediction_mode=prediction_mode)
-        torchmetrics.classification.MultilabelRecall.__init__(self, *args, **kwargs)
+class MulticlassRecall(MulticlassMixin, torchmetrics.classification.MulticlassRecall):
+    pass
 
 
-class MultilabelF1Score(MultilabelExtend, torchmetrics.classification.MultilabelF1Score):
-    def __init__(self, top_k=None, prediction_mode='prob', *args, **kwargs):
-        MultilabelExtend.__init__(self, top_k=top_k, prediction_mode=prediction_mode)
-        torchmetrics.classification.MultilabelF1Score.__init__(self, *args, **kwargs)
+class MulticlassF1Score(MulticlassMixin, torchmetrics.classification.MulticlassF1Score):
+    pass
+
+
+class MultilabelPrecision(MultilabelMixin, torchmetrics.classification.MultilabelPrecision):
+    pass
+
+
+class MultilabelRecall(MultilabelMixin, torchmetrics.classification.MultilabelRecall):
+    pass
+
+
+class MultilabelF1Score(MultilabelMixin, torchmetrics.classification.MultilabelF1Score):
+    pass
